message is now displayed when the browser doesn't support client side storage.

+++ js/storage/db.js
@@ -52,7 +52,9 @@ function () {
      * Deletes the database.
      */
     p.deleteDatabase = function () {
-        window.indexedDB.deleteDatabase(DB_NAME);
+        if (this.supported) {
+            window.indexedDB.deleteDatabase(DB_NAME);
+        }
     };
 
     /**
